corection de la pagination dans un forum

+++ src/com/sdzee/forums/beans/TopicsBackingBean.java
@@ -26,7 +26,8 @@ import com.sdzee.forums.entities.Topic;
 import com.sdzee.membres.entities.Member;
 
 /**
- * TopicsBackingBean est le bean sur lequel s'appuie notamment la page de chaque forum. Il s'agit d'un ManagedBean JSF, ayant pour portée une vue.
+ * TopicsBackingBean est le bean sur lequel s'appuie notamment la page de chaque forum. Il s'agit d'un ManagedBean JSF, ayant pour portée
+ * une vue.
  * 
  * @author Médéric Munier
  * @version %I%, %G%
@@ -36,7 +37,7 @@ import com.sdzee.membres.entities.Member;
 public class TopicsBackingBean implements Serializable {
     private static final long   serialVersionUID   = 1L;
     private static final String URL_TOPIC_PAGE     = "/topic.jsf?topicId=";
-    private static final int    NB_TOPICS_PER_PAGE = 5;
+    private static final double NB_TOPICS_PER_PAGE = 5;
 
     private Topic               topic;
     private Post                post;
@@ -55,17 +56,17 @@ public class TopicsBackingBean implements Serializable {
     private PostDao             postDao;
 
     /**
-     * Cette méthode initialise la variable d'instance <code>forum</code> en récupérant en base le forum correspondant à l'id transmis par la Facelet
-     * <code>forum.xhtml</code>, contenu dans la variable <code>forumId</code>.
+     * Cette méthode initialise la variable d'instance <code>forum</code> en récupérant en base le forum correspondant à l'id transmis par
+     * la Facelet <code>forum.xhtml</code>, contenu dans la variable <code>forumId</code>.
      * <p>
-     * Elle est exécutée automatiquement par JSF, après le constructeur de la classe s'il existe. À l'appel du constructeur classique, le bean n'est
-     * pas encore initialisé, et donc aucune dépendance n'est injectée. Cependant lorsque cette méthode est appelée, le bean est déjà initialisé et il
-     * est donc possible de faire appel à des dépendances. Ici, c'est le DAO {@link ForumDao} injecté via l'annotation <code>@EJB</code> qui entre en
-     * jeu.
+     * Elle est exécutée automatiquement par JSF, après le constructeur de la classe s'il existe. À l'appel du constructeur classique, le
+     * bean n'est pas encore initialisé, et donc aucune dépendance n'est injectée. Cependant lorsque cette méthode est appelée, le bean est
+     * déjà initialisé et il est donc possible de faire appel à des dépendances. Ici, c'est le DAO {@link ForumDao} injecté via l'annotation
+     * <code>@EJB</code> qui entre en jeu.
      * <p>
-     * À la différence de la plupart des autres backing-beans, cette méthode n'est pas annotée avec <code>@PostConstruct</code>. Ceci est simplement
-     * dû au fait qu'elle fait appel à une variable qui est initialisée depuis la vue, en l'occurrence l'id du sujet courant. Puisqu'elle dépend de
-     * l'action du visiteur, son cycle de vie ne peut pas être entièrement géré par JSF.
+     * À la différence de la plupart des autres backing-beans, cette méthode n'est pas annotée avec <code>@PostConstruct</code>. Ceci est
+     * simplement dû au fait qu'elle fait appel à une variable qui est initialisée depuis la vue, en l'occurrence l'id du sujet courant.
+     * Puisqu'elle dépend de l'action du visiteur, son cycle de vie ne peut pas être entièrement géré par JSF.
      */
     public void init() {
         if ( !FacesContext.getCurrentInstance().isPostback() ) {
@@ -74,7 +75,7 @@ public class TopicsBackingBean implements Serializable {
             forum = forumDao.find( forumId );
             pagesNumber = (int) Math.ceil( topicDao.count( forum ) / NB_TOPICS_PER_PAGE );
             stickyTopics = topicDao.listStickies( forum );
-            paginatedTopics = topicDao.list( forum, page, NB_TOPICS_PER_PAGE );
+            paginatedTopics = topicDao.list( forum, page, (int) NB_TOPICS_PER_PAGE );
         }
     }
